Rename sysinfo frame to be researchable
