+++ src/browser/modules/Stream/SysInfoFrame/SysInfoFrame.tsx
